+++ src/Components/mainPage/Keyboard.js
@@ -1,106 +1,187 @@
 import React, { Component } from 'react'
 import "./Keyboard.css"
-import { getPitchChanger, delayMs } from "../SongUtils"
-import Instrument from "../Instrument"
+import { delayMs, ComposerToRecording, ApproachingSong } from "../SongUtils"
 import Note from "./Note"
-import { keyNames, pitchArr , layoutImages, appName,layoutData} from "../../appConfig"
-import {songStore} from './SongStore'
+import { keyNames, pitchArr, layoutImages, appName, layoutData } from "../../appConfig"
+import { songStore } from './SongStore'
 import { observe } from 'mobx'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { faSyncAlt, faStop } from '@fortawesome/free-solid-svg-icons'
 class Keyboard extends Component {
     constructor(props) {
         super(props)
+        let propKeyboard = this.props.data.keyboard
+        propKeyboard.layout.forEach(note => { note.status = '' })
+
         this.state = {
-            instrument: new Instrument(),
-            audioContext: new (window.AudioContext || window.webkitAudioContext)(),
-            reverbAudioContext: new (window.AudioContext || window.webkitAudioContext)(),
             playTimestamp: new Date().getTime(),
             songToPractice: [],
+            approachingNotes: ApproachingSong(appName === 'Sky' ? 15 : 21),
+            keyboard: propKeyboard.layout,
             sliderState: {
                 position: 0,
                 size: 0
+            },
+            approachingScore: {
+                correct: 1,
+                wrong: 1
             }
         }
-        console.log("Loaded:",props.data.instrument)
-        this.loadInstrument(props.data.instrument)
+        this.approachRate = 1500    
+        this.approachingNotesList = []
+        this.songTimestamp = 0
+        this.disposeStore = () => { }
+        this.tickTime = 50
+        this.tickInterval = 0
+    }
 
-        try {
-            this.loadReverb()
-        } catch(e) {
-            console.log("Error with reverb0",e)
-        }
-    }
-    handleKeyboard = (event) => {
-        if(event.repeat) return
-        if (document.activeElement.tagName === "INPUT") return
-        let code = event.code?.replace("Key","")
-        let index = this.state.instrument.getNoteFromCode(code)
-        let note
-        if (index !== null) {
-            note = this.state.instrument.layout[index]
-        }
-        if (note !== undefined) {
-            this.handleClick(note)
-        }
-    }
     componentDidMount() {
         window.addEventListener('keydown', this.handleKeyboard)
-        this.disposeStore = observe(songStore,(data) => {
+        this.tickInterval = setInterval(this.tick,this.tickTime)
+        this.disposeStore = observe(songStore,async (data) => {
             let value = data.object.data
             let type = value.eventType
             let lostReference = JSON.parse(JSON.stringify(value.song))
+            if (lostReference.data?.isComposedVersion) {
+                lostReference = ComposerToRecording(lostReference)
+            }
             lostReference.timestamp = new Date().getTime()
-            if(type === 'play') this.handlePlayEvent(lostReference)
-            if(type === 'practice') this.handlePracticeEvent(lostReference)
+            let hasSong = false
+            if (type === 'play') {
+                await this.stopSong()
+                this.playSong(lostReference, value.start)
+                hasSong = true
+            }
+            if (type === 'practice') {
+                await this.stopSong()
+                this.practiceSong(lostReference, value.start)
+                hasSong = true
+            }
+            if(type === 'approaching'){
+                await this.stopSong()
+                this.approachingSong(lostReference, value.start)
+                hasSong = true
+            }
+            if (type === 'stop') await this.stopSong()
+            this.props.functions.setHasSong(hasSong)
         })
     }
     componentWillUnmount() {
         window.removeEventListener('keydown', this.handleKeyboard)
         this.disposeStore()
-        let state = this.state
-        state.playTimestamp = new Date().getTime()
+        this.songTimestamp = 0
+        clearInterval(this.tickInterval)
     }
-    loadReverb() {
-        let audioCtx = this.state.audioContext
-        fetch("./assets/audio/reverb4.wav")
-            .then(r => r.arrayBuffer())
-            .then(b => {
-                audioCtx.decodeAudioData(b, (impulse_response) => {
-                    let convolver = audioCtx.createConvolver()
-                    let gainNode = audioCtx.createGain()
-                    gainNode.gain.value = 2.5
-                    convolver.buffer = impulse_response
-                    convolver.connect(gainNode)
-                    gainNode.connect(audioCtx.destination)
-                    this.setState({
-                        reverbAudioContext: convolver
-                    })
-                })
-            }).catch((e) => { 
-                console.log("Error with reverb1",e)
-            })
-    }
-    handlePlayEvent = (data) => {
-        this.setState({
-            playTimestamp: data.timestamp
-        }, () => this.playSong(data))
-    }
-    handlePracticeEvent = (data) => {
-        this.practiceSong(data)
-    }
-    loadInstrument = async (name) => {
-        let newInstrument = new Instrument(name)
-        await newInstrument.load(this.state.audioContext)
 
-        this.setState({
-            instrument: newInstrument
+    approachingSong = async (song) => {
+        let notes = []
+        this.approachRate = this.props.data.approachRate || 1500
+        let startDelay = this.approachRate
+        song.notes.forEach(note => {
+            let obj = {
+                time: note[1] + startDelay,
+                index: note[0]
+            }
+            notes.push(obj)
         })
-
+        this.setState({
+            approachingNotes: ApproachingSong(appName === 'Sky' ? 15 : 21),
+            approachingScore: {
+                wrong: 1,
+                correct: 1
+            }
+        })
+        this.changeSliderState(0,notes.length)
+        this.approachingNotesList = notes
     }
-    practiceSong = (song) => {
+
+    tick = () => {
+        if(!this.props.data.hasSong) return
+        const { approachingNotes, sliderState , approachingScore} = this.state
+        let stateNotes = approachingNotes
+        let notes = this.approachingNotesList
+        notes.forEach(note => {
+            note.time -= this.tickTime
+        })
+        let hasChanges = false
+        let removed = 0
+        for(let i = 0; i< notes.length; i++){
+            if(notes[i].time < this.approachRate){
+                stateNotes[notes[i].index].push({
+                    clicked: false,
+                    time: this.approachRate,
+                    id: Math.floor(Math.random() * 10000)
+                })
+                notes.splice(i,1)
+                i--
+                hasChanges = true
+            }else{
+                break
+            }
+        }
+        stateNotes.forEach(note => {
+            for(let i = 0; i< note.length; i++){
+                let apNote = note[i]
+                apNote.time -= this.tickTime
+                if(apNote.clicked){ 
+                    if(apNote.time < this.approachRate / 3){
+                        approachingScore.correct++
+                    }else{
+                        approachingScore.wrong++
+                    }
+                    apNote.time = -1 //so that it can be removed after
+                }
+                if(apNote.time < 0){
+                    if(!apNote.clicked) approachingScore.wrong++
+                    note.splice(i,1)
+                    i--
+                    hasChanges = true
+                    removed ++
+                }
+            }
+        })
+        if(!hasChanges) return
+        this.changeSliderState(sliderState.position + removed)  
+        this.setState({
+            approachingNotes: stateNotes,
+            approachingScore: approachingScore
+        })
+    }
+
+    playSong = async (song, start = 0) => {
+        this.songTimestamp = song.timestamp
+        const { keyboard } = this.state
+        let notes = song.notes
+        if (notes.length === 0) return
+        let previous = notes[0][1]
+        let pastError = 0
+        let previousTime = new Date().getTime()
+        for (let i = start; i < notes.length; i++) {
+            let delay = notes[i][1] - previous
+            previous = notes[i][1]
+            let note = notes[i][0]
+            if (this.songTimestamp !== song.timestamp) break
+            previousTime = new Date().getTime()
+            if (delay > 16) await delayMs(delay - pastError)
+            keyboard[note].status = 'clicked'
+            this.handleClick(keyboard[note])
+            this.setState({
+                keyboard,
+                sliderState: {
+                    position: i,
+                    size: notes.length
+                }
+            })
+            pastError = new Date().getTime() - previousTime - delay
+        }
+        songStore.data = returnStopSong()
+    }
+
+    practiceSong = (song, start = 0) => {
+        const { keyboard } = this.state
         let notes = song.notes
         let songLength = notes.length
-        if (song.start === undefined) song.start = 0
-        notes.splice(0, song.start)
+        notes.splice(0, start)
         let chunks = []
         for (let i = 0; notes.length > 0; i++) {
             let chunk = {
@@ -109,7 +190,7 @@ class Keyboard extends Component {
             }
             let startTime = chunk.notes.length > 0 ? chunk.notes[0][1] : 0
             for (let j = 0; j < notes.length && j < 20; j++) {
-                let difference = notes[j][1] - chunk.notes[0][1] - song.threshold
+                let difference = notes[j][1] - chunk.notes[0][1] - 50 //TODO add threshold here
                 if (difference < 0) {
                     chunk.notes.push(notes.shift())
                     j--
@@ -118,133 +199,212 @@ class Keyboard extends Component {
             chunk.delay = notes.length > 0 ? notes[0][1] - startTime : 0
             chunks.push(chunk)
         }
-        this.setState({
-            songToPractice: chunks
+        if (chunks.length === 0) return
+        let firstChunk = chunks[0]
+        let secondChunk = chunks[1]
+        firstChunk.notes.forEach(note => { keyboard[note[0]].status = 'toClick' })
+        secondChunk?.notes.forEach(note => {
+            let keyboardNote = keyboard[note[0]]
+            if (keyboardNote.status === 'toClick') return keyboardNote.status = 'toClickAndNext'
+            keyboardNote.status = 'toClickNext'
         })
-        this.setSlider({
-            size: songLength,
-            position: song.start
+        this.props.functions.setHasSong(true)
+        this.setState({
+            songToPractice: chunks,
+            keyboard,
+            sliderState: {
+                position: start,
+                size: songLength
+            }
         })
     }
-    changeSliderPosition = (position) => {
-        let sliderState = this.state.sliderState
-        sliderState.position += position
-        this.setState({
-            sliderState: sliderState
-        }, () => this.props.functions.changeSliderState(this.state.sliderState))
+    restartSong = () => {
+        let lostReference = JSON.parse(JSON.stringify(songStore.data))
+        this.stopSong()
+        setTimeout(() => {
+            let start = songStore.data.start
+            if(songStore.data.eventType === 'practice')
+                start = this.state.sliderState.position
+            songStore.data = {
+                ...lostReference,
+                start
+            }
+        }, 400)
     }
-    setSlider = (state) => {
-        this.setState({
-            sliderState: state
-        }, this.props.functions.changeSliderState(state))
-    }
-    playSong = async (song) => {
-        let notes = song.notes
-        let previous = 0
-        this.setSlider({
-            size: notes.length,
-            position: 0
+    stopSong = () => {
+        return new Promise(res => {
+            this.songTimestamp = 0
+            const { keyboard } = this.state
+            keyboard.forEach(note => { note.status = '' })
+            this.approachingNotesList = []
+            this.setState({ 
+                keyboard,
+                songToPractice: [],
+                approachingNotes: ApproachingSong(appName === 'Sky' ? 15 : 21)
+            },res)
+
+            this.props.functions.setHasSong(false)
         })
-        if (notes.length === 0) return
-        for (let i = 0; i < notes.length; i++) {
-            let delay = notes[i][1] - previous
-            previous = notes[i][1]
-            let note = notes[i][0]
-            if (this.state.playTimestamp !== song.timestamp) break
-            if (delay > 10) await delayMs(delay)
-            this.changeSliderPosition(1)
-            this.playSound(this.state.instrument.layout[note])
+
+    }
+    handleSliderEvent = (event) => {
+        this.changeSliderState(Number(event.target.value))
+    }
+    changeSliderState = async (position, size = this.state.sliderState.size) => {
+        this.setState({
+            sliderState: {
+                position,
+                size
+            }
+        })
+    }
+    handleKeyboard = (event) => {
+        const { keyboard } = this.state
+        if (event.repeat) return
+        if (document.activeElement.tagName === "INPUT") return
+        let code = event.code?.replace("Key", "")
+        let index = this.props.data.keyboard.getNoteFromCode(code)
+        let note = keyboard[index]
+        if (note) this.handleClick(note)
+    }
+
+    handleApproachClick = (note) => {
+        const { approachingNotes } = this.state
+        let approachingNote = approachingNotes[note.index][0]
+        if(approachingNote){
+            approachingNote.clicked = true
         }
-        this.props.functions.stopSong()
     }
-    handleClick = (note) => {
-        let practiceSong = this.state.songToPractice
-        if (practiceSong.length > 0) {
-            let indexClicked = practiceSong[0]?.notes.findIndex(e => e[0] === note.index)
+    handlePracticeClick = (note) => {
+        const { keyboard, songToPractice, sliderState } = this.state
+        if (songToPractice.length > 0) {
+            let indexClicked = songToPractice[0]?.notes.findIndex(e => e[0] === note.index)
             if (indexClicked !== -1) {
-                practiceSong[0].notes.splice(indexClicked, 1)
-                if (practiceSong[0].notes.length === 0) practiceSong.shift()
-                if (practiceSong.length === 0) this.props.functions.stopSong()
+                songToPractice[0].notes.splice(indexClicked, 1)
+                if (songToPractice[0].notes.length === 0) songToPractice.shift()
+                if (songToPractice.length === 0) {
+                    this.stopSong()
+                    songStore.data = returnStopSong()
+                } else {
+                    let nextChunk = songToPractice[0]
+                    let nextNextChunk = songToPractice[1]
+                    nextChunk.notes.forEach(note => keyboard[note[0]].status = 'toClick')
+                    nextNextChunk?.notes.forEach(note => {
+                        let keyboardNote = keyboard[note[0]]
+                        if (keyboardNote.status === 'toClick') return keyboardNote.status = 'toClickAndNext'
+                        keyboardNote.status = 'toClickNext'
+                    })
+                }
                 this.setState({
-                    songToPractice: practiceSong
-                }, () => this.changeSliderPosition(1))
+                    songToPractice: songToPractice,
+                    sliderState: {
+                        size: sliderState.size,
+                        position: sliderState.position + 1
+                    }
+                })
             }
         }
-
-        this.playSound(note)
-
     }
-    playSound = (note) => {
-        const {state, props} = this
-        if(note === undefined) return
-        if (props.isRecording) props.functions.handleRecording(note)
-        note.clicked = true
+    handleClick = (note, onlySound) => {
+        const { keyboard } = this.state
+        keyboard[note.index].status = 'clicked'
+        this.handlePracticeClick(note)
+        this.handleApproachClick(note)
+        this.setState({ keyboard })
         setTimeout(() => {
-            note.clicked = false
-            this.setState({
-                instrument: state.instrument
-            })
+            if (note.status !== 'clicked') return
+            keyboard[note.index].status = ''
+            this.setState({ keyboard })
         }, 200)
-        const source = state.audioContext.createBufferSource()
-        source.playbackRate.value = getPitchChanger(props.data.pitch)
-        source.buffer = note.buffer
-        if (props.data.caveMode) {
-            source.connect(state.reverbAudioContext)
-        } else {
-            source.connect(state.audioContext.destination)
-        }
-        source.start(0)
-        this.setState({
-            instrument: state.instrument
-        })
+        this.props.functions.playSound(note)
     }
     render() {
-        const {state, props} = this
-        const {data} = props
+        const { state, props } = this
+        const { data } = props
+        const { keyboard, approachingScore } = state
         let size = data.keyboardSize / 100
         if (size < 0.5) size = 0.5
         if (size > 1.5) size = 1.5
         let keyboardClass = "keyboard"
-        if (state.instrument.layout.length === 15) keyboardClass += " keyboard-5"
-        if (state.instrument.layout.length === 8) keyboardClass += " keyboard-4"
-        return <div 
-                    className={keyboardClass} 
-                    style={{ 
-                        transform: `scale(${size})`,
-                        marginBottom:size * 30
-                    }}
-                >
-             {state.instrument.layout.length === 0 ? <div className="loading">Loading...</div> : null}
-            {state.instrument.layout.map(note => {
-                let toBeClicked = state.songToPractice[0]?.notes.find(e => e[0] === note.index) !== undefined
-                let toBeClickedNext = state.songToPractice[1]?.notes.find(e => e[0] === note.index) !== undefined
-                let fadeTime = state.songToPractice[0]?.delay !== undefined ? state.songToPractice[0]?.delay / 1000 : 0.1
-                let noteText = ""
-                let noteImage = ""
-                try{
-                    noteImage = layoutImages[state.instrument.layout.length][note.index]
-                    if(data.noteNameType === "Note name") noteText = keyNames[appName][pitchArr.indexOf(data.pitch)][note.index]                           
-                    if(data.noteNameType === "Keyboard layout") noteText = layoutData[state.instrument.layout.length].keyboardLayout[note.index]
-                    if(data.noteNameType === "Do Re Mi") noteText = layoutData[state.instrument.layout.length].mobileLayout[note.index]
-                }catch(e){}
+        if (keyboard.length === 15) keyboardClass += " keyboard-5"
+        if (keyboard.length === 8) keyboardClass += " keyboard-4"
+        return <>   
+            {data.hasSong &&
+                <div className="upper-right">
+                    <div className='approaching-accuracy'>
+                        Accuracy: %{(approachingScore.correct / ( approachingScore.correct + approachingScore.wrong - 1) * 100 ).toFixed(1)}
+                    </div>
+                    <div className="slider-wrapper">
+                        <button className="song-button" onClick={() => {
+                            this.stopSong()
+                            songStore.data = returnStopSong()
+                        }
+                        }>
+                            <FontAwesomeIcon icon={faStop} />
+                        </button>
+                        <input
+                            type="range"
+                            className="slider"
+                            min={0}
+                            onChange={this.handleSliderEvent}
+                            max={state.sliderState.size}
+                            value={state.sliderState.position}
+                        ></input>
+                        <button className="song-button" onClick={() => {
+                            this.restartSong()
+                        }}>
+                            <FontAwesomeIcon icon={faSyncAlt} />
+                        </button>
+                    </div>
+                </div>
+            }
+            <div
+                className={keyboardClass}
+                style={{
+                    transform: `scale(${size})`,
+                    marginBottom: size * 30
+                }}
+            >
+                {keyboard.length === 0 ? <div className="loading">Loading...</div> : null}
+                {keyboard.map(note => {
+                    let noteImage = layoutImages[keyboard.length][note.index]
+                    let approachingNotes = state.approachingNotes[note.index]
+                   
+                    return <Note
+                        key={note.index}
+                        fadeTime={0}
+                        data={note}
+                        status={note.status}
+                        handleClick={this.handleClick}
+                        approachingNotes={approachingNotes}
+                        approachRate={this.approachRate}
+                        instrument={this.props.data.keyboard.instrumentName}
+                        noteText={getNoteText(data.noteNameType, note.index, data.pitch, keyboard.length)}
+                        noteImage={`./assets/icons/keys/${noteImage}.svg`}
+                    />
 
-                return <Note
-                    key={note.index}
-                    toBeClicked={toBeClicked}
-                    fadeTime={fadeTime}
-                    toBeClickedNext={toBeClickedNext}
-                    data={note}
-                    noteText={noteText}
-                    instrument={state.instrument.instrumentName}
-                    noteImage={`./assets/icons/keys/${noteImage}.svg`}
-                    clickAction={this.handleClick}
-                >
+                })}
+            </div>
+        </>
 
-                </Note>
-            })}
-        </div>
+
+    }
+}
+function returnStopSong() {
+    return {
+        song: {},
+        eventType: 'stop',
+        start: 0
     }
 }
 
+function getNoteText(noteNameType, index, pitch, layoutLength) {
+    try {
+        if (noteNameType === "Note name") return keyNames[appName][pitchArr.indexOf(pitch)][index]
+        if (noteNameType === "Keyboard layout") return layoutData[layoutLength].keyboardLayout[index]
+        if (noteNameType === "Do Re Mi") return layoutData[layoutLength].mobileLayout[index]
+    } catch (e) {}
+    return ''
+}
 export default Keyboard
