+++ src/service-worker.js
@@ -7,7 +7,7 @@
 // You can also remove this file if you'd prefer not to use a
 // service worker, and the Workbox build step will be skipped.
 import { appName } from './appConfig';
-const CACHE = appName+"-3.1"
+const CACHE = appName+"-3.2"
 console.log("CACHE Version: ",CACHE)
 import { clientsClaim } from 'workbox-core';
 import { ExpirationPlugin } from 'workbox-expiration';
